#38 progress - Shared Credential page now allows saving of a Private Key value.
A 'credential name' is required.  It's assumed the feature in the Task
Engine will involve the New Connection command referencing the
credential by it's name.

The 'Asset' box is currently a sort of free-for-all entry box on the
New Connection command.  It's parsed for specific content, for an asset
id/name, for a connection string, as well as a shared_credential.

I updated catocommon.lookup_shared_cred  to return the pk.  See
catotaskengine/commands.py line 1132 and 1210 for the references.
However, it seems there are several different methods hitting the
asset_credential table (lookup_shared_cred, gather_system_info) so I
stopped … be careful when pulling those threads, and consolidate if
possible while you're in there.

Note: the Assed Edit page has a terrible, hackish way of creating a
credential.  You cannot at this time create a shared credential with a
PK from inside the Asset Edit dialog, nor can a credential explicitly
saved on an asset be a PK type.